backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12
python-dotenv>=1.0.0

# LangGraph / LangChain
langgraph>=0.2.50
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-mcp-adapters>=0.1.0

# PDF parsing
pdfplumber>=0.11.0

# Pydantic
pydantic>=2.0.0
